+++ NovaEngine/Content/Readers/FontReader.cs
@@ -32,7 +32,7 @@ public class FontReader : IContentReader
         {
             var character = binaryReader.ReadChar();
             var size = new Vector2I(binaryReader.ReadUInt16(), binaryReader.ReadUInt16());
-            var atlasPosition = new Rectangle(binaryReader.ReadUInt16(), binaryReader.ReadUInt16(), binaryReader.ReadUInt16(), binaryReader.ReadUInt16());
+            var atlasPosition = new Rectangle(binaryReader.ReadSingle(), binaryReader.ReadSingle(), binaryReader.ReadSingle(), binaryReader.ReadSingle());
             var horizontalMetrics = new HorizontalMetrics(binaryReader.ReadUInt16(), binaryReader.ReadUInt16());
 
             glyphs.Add(new(character, size, atlasPosition, horizontalMetrics));
